% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_language_and_path.R
\name{audio_language_and_path}
\alias{audio_language_and_path}
\title{Generate a small data set with the values of \code{language}, \code{language_code} and \code{audio_path}}
\usage{
audio_language_and_path()
}
\description{
This information is important to download and check the audio files easily using loops.
}
