% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audios_download_check.R
\name{audios_download_check}
\alias{audios_download_check}
\title{Download audios for the words in the data set and check missing or unused audios}
\usage{
audios_download_check(language)
}
\description{
\enumerate{
\item Download audios and store them in the appropriate directory.
\item Create a text file with the missing audios.
\item Remove unused audios.
The function \code{download_audios_all()} loops over all the available languages in
\code{\link[=audio_language_and_path]{audio_language_and_path()}}.
}
}
