% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_audios.R
\name{check_audios}
\alias{check_audios}
\title{Check the audio files from a list of words}
\usage{
check_audios(data, language, audio_dir)
}
\arguments{
\item{data}{data set with the data}

\item{language}{}

\item{audio_dir}{path to the directory of the audio files}
}
\value{
A list with 2 elements indicating the problems, if any, between the data set and
the audio files.
}
\description{
Check the audio files from a list of words
}
