% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_images.R
\name{check_images}
\alias{check_images}
\title{Check that different parameters are correct in the project}
\usage{
check_images(data_set, image_folder)
}
\arguments{
\item{data_set}{path to the data set}

\item{image_folder}{path to the folder of the images}
}
\value{
A list with different results, everything should be TRUE if there are no problems.
}
\description{
Check some things to be sure that the project does not have errors, mostly related to images.
}
