% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_audios.R
\name{download_audios}
\alias{download_audios}
\title{Download audios for the words in the data set}
\usage{
download_audios(language, language_code, dir_media)
}
\description{
This functions is used to download audios and store them in the appropriate directory.
It also creates a text file with the missing words (words without available audios).
}
