% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_output.R
\name{generate_output}
\alias{generate_output}
\title{Create output files for a language in an specified directory}
\usage{
generate_output(language, path)
}
\value{

}
\description{
Use the main data set to generate a data set to be imported to Anki, it contains the
information for one language, and it generates columns with HTML format (e.g. audios and
images).
}
