% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tts_sound_of_text.R
\name{tts_sound_of_text}
\alias{tts_sound_of_text}
\title{Download audio files from text using the Sound of Text API}
\usage{
tts_sound_of_text(words, voice, language, folder_path, sleep_time = 2)
}
\arguments{
\item{words}{string or vector of strings}

\item{voice}{voice for generating audio files, check https://soundoftext.com/docs#voices}

\item{language}{language of the words (used for naming the columns)}

\item{folder_path}{path of the folder that will contain the MP3 files of the words}

\item{sleep_time}{time to sleep between downloading each word}
}
\value{

}
\description{
Text to speech function to download MP3 files using an API of Sound of Text
https://soundoftext.com/
}
\examples{
tts_sound_of_text(c("hello", "world"), "en-US", "data/audios", "vocab-french-")
}
