% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_output.R
\name{generate_output}
\alias{generate_output}
\title{Generate output data set ready to be imported to Anki}
\usage{
generate_output(language)
}
\arguments{
\item{language}{The language which is used to generate the data.}

\item{data}{The main data set.}
}
\value{
A tibble.
}
\description{
Use the main data set to generate a data set to be imported to Anki, it contains the
information for one language, and it generates columns with HTML format (e.g. audios and
images).
}
\examples{
generate_output("spanish")
}
